+++ vtm/src/org/oscim/layers/LocationTextureLayer.java
@@ -1,6 +1,8 @@
 /*
+ * Copyright 2013 Ahmad Saleem
+ * Copyright 2013 Hannes Janetzek
+ * Copyright 2016-2019 devemux86
  * Copyright 2017-2018 Longri
- * Copyright 2018 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -15,19 +17,22 @@
  */
 package org.oscim.layers;
 
+import org.oscim.backend.CanvasAdapter;
 import org.oscim.core.MercatorProjection;
 import org.oscim.map.Map;
 import org.oscim.renderer.LocationTextureRenderer;
-import org.oscim.renderer.atlas.TextureRegion;
 
 public class LocationTextureLayer extends Layer {
     public final LocationTextureRenderer locationRenderer;
 
-    public LocationTextureLayer(Map map, TextureRegion textureRegion) {
+    public LocationTextureLayer(Map map) {
+        this(map, CanvasAdapter.getScale());
+    }
+
+    public LocationTextureLayer(Map map, float scale) {
         super(map);
 
-        mRenderer = locationRenderer = new LocationTextureRenderer(map);
-        locationRenderer.setTextureRegion(textureRegion);
+        mRenderer = locationRenderer = new LocationTextureRenderer(map, this, scale);
     }
 
     @Override
@@ -41,14 +46,11 @@ public class LocationTextureLayer extends Layer {
             locationRenderer.animate(false);
     }
 
-    public void setPosition(double latitude, double longitude, float bearing, float accuracy) {
+    public void setPosition(double latitude, double longitude, float accuracy) {
         double x = MercatorProjection.longitudeToX(longitude);
         double y = MercatorProjection.latitudeToY(latitude);
-        bearing = -bearing;
-        while (bearing < 0)
-            bearing += 360;
         double radius = accuracy / MercatorProjection.groundResolutionWithScale(latitude, 1);
-        locationRenderer.setLocation(x, y, bearing, radius);
+        locationRenderer.setLocation(x, y, radius);
         locationRenderer.animate(true);
     }
 }