+++ docs/Changelog.md
@@ -3,6 +3,7 @@
 ## New since 0.12.0
 
 - Render themes: symbols on lines with billboard / rotation [#743](https://github.com/mapsforge/vtm/pull/743)
+- Location texture renderer: rewrite and optimize [#750](https://github.com/mapsforge/vtm/pull/750)
 - Android: OpenGL ES 2.0 default for performance / stability [#749](https://github.com/mapsforge/vtm/pull/749)
   - `MapView.OPENGL_VERSION`
 - Android: threaded system initialization